rename some variables, file transfer
