+++ lib/configuration files/blocksSettings.dart
